Add TweetDetailActivity to show details of a Tweet
When a user clicks on a tweet in Timeline, the TweetDetailActivity shows
up. In later commits, I will add different widgets in there, but for now
it only shows the title "Tweet".

Centering the Title of the activity takes some work.

I may want to use a ViewPager later so users can swipe left and right to
look at details of other tweets in TweetDetailActivity directly.