A user can like/unlike a tweet at detail view
The like/unlike causes the newest fetched ID to be the min of (newest
fetched id, the tweet id). It's to ensure the tweet gets a fresh copy of
the like counts after the like/unlike action.

While the detail view can show the like/unlike icon immediately, not
sure how to do it in timeline (currently not showing the like counts).