A user can retweet/unretweet a tweet at detail view
The retweet/unretweet causes the newest fetched ID to be the min of (newest
fetched id, the tweet id). It's to ensure the tweet gets a fresh copy of
the retweet counts after the retweet/unretweet action.

While the detail view can show the retweet/unretweet icon immediately, not
sure how to do it in timeline (currently not showing the retweet counts).