Fix retweeted and liked Tweet is not reflected on timeline
When a user retweets or likes a Tweet in detail view, going back to
timeline and go to the detail view again does not see the state being
updated. It's because the cursor of the TimelineActivity was not changed
by the TweetDetailFragment. Fix this by using startActivityForResult and
onActivityResult in TimelineActivity.