Create UserTweet model, fix fetched id in TweetDetailFragment
The UserTweet model will be like Tweet and Mention.

Move UserTimelineFragment to the package timeline, so it can use the
protected memebers in TimelineFragment. Implemented the fetchNewerTweets
method properly in there as well.

Fix fetched id in TweetDetailFragment by passing in the field that we
use to fetch newest id from SimpleTweetPrefs.