Finish Followers screen
Complete the generic implementation of UsersFragment, FollowersFragment
can extend it to just fetch the followers

Clicking on Followers from Profile can see the list of followers of the
authenticated user. It cannot be done for arbitrary users because the
API can only fetch a list of user id's for the followers of an arbitrary
users.